Generalised Zip flowgraph nodes: now have ZipN and ZipSet
